# api/openapi.yaml
openapi: 3.0.3
info:
  title: Kafkas Derneği Panel API
  description: Kafkas Derneği yönetim paneli API dokümantasyonu
  version: 1.0.0
  contact:
    name: API Support
    email: support@kafkasder.org
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.kafkasder.org/api
    description: Production server
...
paths:
...
  /members:
    get:
      tags:
        - Members
      summary: Üye listesi
      description: Tüm üyeleri sayfalı olarak getirir
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
            description: Ad, soyad veya TC Kimlik No ile arama
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '500':
          description: Sunucu hatası
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Members
      summary: Yeni üye oluştur
      description: Sisteme yeni üye ekler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInput'
      responses:
        '201':
          description: Üye başarıyla oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Geçersiz veri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /members/{id}:
    get:
      tags:
        - Members
      summary: Üye detayı
      description: Belirli bir üyenin detaylarını getirir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          description: Üye bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Members
      summary: Üye güncelle
      description: Belirli bir üyeyi günceller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdateInput'
      responses:
        '200':
          description: Başarılı güncelleme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Geçersiz veri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Üye bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Members
      summary: Üye sil
      description: Belirli bir üyeyi siler
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Başarılı silme
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Üye başarıyla silindi
        '500':
          description: Sunucu hatası
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /social-aid/{id}:
    get:
      tags:
        - Social Aid
      summary: Başvuru detayı
      description: Belirli bir başvurunun detaylarını getirir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialAidApplication'
        '404':
          description: Başvuru bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Social Aid
      summary: Başvuru güncelle
      description: Belirli bir başvuruyu günceller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialAidApplicationUpdateInput'
      responses:
        '200':
          description: Başarılı güncelleme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialAidApplication'
        '400':
          description: Geçersiz veri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Başvuru bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /settings:
    get:
      tags:
        - Settings
      summary: Ayar listesi
      description: Tüm ayarları getirir
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Setting'
        '500':
          description: Sunucu hatası
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Settings
      summary: Yeni ayar oluştur
      description: Sisteme yeni ayar ekler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingInput'
      responses:
        '201':
          description: Ayar başarıyla oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
        '400':
          description: Geçersiz veri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /settings/{id}:
    get:
      tags:
        - Settings
      summary: Ayar detayı
      description: Belirli bir ayarın detaylarını getirir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
        '404':
          description: Ayar bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Settings
      summary: Ayar güncelle
      description: Belirli bir ayarı günceller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingUpdateInput'
      responses:
        '200':
          description: Başarılı güncelleme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
        '400':
          description: Geçersiz veri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ayar bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Settings
      summary: Ayar sil
      description: Belirli bir ayarı siler
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Başarılı silme
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Ayar başarıyla silindi
        '500':
          description: Sunucu hatası
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Hata mesajı
        message:
          type: string
          example: Detaylı hata açıklaması
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    MemberInput:
      type: object
      required:
        - tc_kimlik_no
        - ad
        - soyad
        - cinsiyet
        - telefon
        - uye_turu
      properties:
        tc_kimlik_no:
          type: string
          minLength: 11
          maxLength: 11
        ad:
          type: string
          minLength: 2
        soyad:
          type: string
          minLength: 2
        dogum_tarihi:
          type: string
          format: date
        cinsiyet:
          type: string
          enum: [erkek, kadin]
        telefon:
          type: string
          minLength: 10
        email:
          type: string
          format: email
        adres:
          type: string
        uye_turu:
          type: string
          enum: [onursal, standart, fahri]
        aidat_durumu:
          type: string
          enum: [odendi, gecikti, beklemede]
        kayit_tarihi:
          type: string
          format: date
    MemberUpdateInput:
      allOf:
        - $ref: '#/components/schemas/MemberInput'
      type: object
    Member:
      allOf:
        - $ref: '#/components/schemas/MemberInput'
      type: object
      properties:
        id:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SocialAidApplicationInput:
      type: object
      required:
        - beneficiary_id
        - yardim_turu
        - gerekce
      properties:
        beneficiary_id:
          type: number
        yardim_turu:
          type: string
          enum: [gida, egitim, saglik, barinma, giyim, diger]
        talep_edilen_tutar:
          type: number
        gerekce:
          type: string
          minLength: 10
        durum:
          type: string
          enum: [beklemede, inceleniyor, onaylandi, reddeddi]
    SocialAidApplicationUpdateInput:
      allOf:
        - $ref: '#/components/schemas/SocialAidApplicationInput'
      type: object
    SocialAidApplication:
      allOf:
        - $ref: '#/components/schemas/SocialAidApplicationInput'
      type: object
      properties:
        id:
          type: number
        basvuru_tarihi:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SettingInput:
      type: object
      required:
        - key
      properties:
        key:
          type: string
        value:
          type: string
        category:
          type: string
          enum: [genel, uyelik, bagis, sosyal-yardim, sistem]
        description:
          type: string
    SettingUpdateInput:
      allOf:
        - $ref: '#/components/schemas/SettingInput'
      type: object
    Setting:
      allOf:
        - $ref: '#/components/schemas/SettingInput'
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time